Split up `Retry-After` and IETF parsing into separate modules
