+++ src/convert.rs
@@ -1,4 +1,4 @@
-use super::Result;
+use crate::error::Result;
 
 pub(crate) fn to_usize(value: &str) -> Result<usize> {
     Ok(value.trim().parse::<usize>()?)
